refactor(direct-control): extract setup read owner
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves only setup snapshot and setup map row read orchestration, command builders, shared setup snapshot/map-row source, and map script validation into packages/civ7-direct-control/src/setup/reads.ts while keeping public facade exports in index.ts.

Preserves setup map script validation, setup map row limit default/bounds, setup-domain/config-db row materialization, setup snapshot phase/config shape, and lifecycle helper reuse of the same setup source. ensureCiv7SetupMapRowVisible, prepare/start/run setup lifecycle, restart/begin orchestration, no-replay ownership, hotseat runtime proof, AI ingestion, semantic CLI projection, telemetry, Effect/oRPC procedure-core work, and Task 2.9.4 matrix-row acceptance remain pending.

Validation:
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test -- setup-and-lifecycle.test.ts
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- git diff --check
- ownership and no-cycle rg scans

Local package/source relocation proof only; no runtime/live-game, lifecycle behavior-change, hotseat product-path, AI-ingestion, semantic CLI, telemetry, or procedure-core proof claimed.
